Oops. Bugfix.
Rev: src/errors.h:1.13

+++ src/errors.h
@@ -3,7 +3,7 @@
 #ifdef ERR_DECLARE
 #define DECLARE_ERROR(NAME, INHERIT, DECL) \
 struct program *PIKE_CONCAT(NAME,_error_program); \
-int PIKE_CONCAT(NAME,_error_offset); 
+ptrdiff_t PIKE_CONCAT(NAME,_error_offset); 
 
 #define ERR_VAR(TYPE,CTYPE,RUNTYPE,NAME) TYPE NAME ;
 #endif
